Style : " GameObjectInfo 수정 "
GameObjectInfo
 - st_StatInfo
  - Level 자료형 수정 int -> byte

+++ Assets/Scripts/Client/GameObject/GameObjectInfo.cs
@@ -1020,7 +1020,7 @@ public class st_PositionInfo : IEquatable<st_PositionInfo>
 
 public class st_StatInfo
 {
-    public int Level;
+    public byte Level;
     public int HP;
     public int MaxHP;
     public int MP;
